+++ lib/dynamics_crm.rb
@@ -10,6 +10,7 @@ require "dynamics_crm/xml/query"
 require "dynamics_crm/xml/fetch_expression"
 require "dynamics_crm/xml/entity"
 require "dynamics_crm/xml/entity_reference"
+require "dynamics_crm/xml/entity_collection"
 require "dynamics_crm/response/result"
 require "dynamics_crm/response/retrieve_result"
 require "dynamics_crm/response/retrieve_multiple_result"
@@ -17,14 +18,25 @@ require "dynamics_crm/response/create_result"
 require "dynamics_crm/response/execute_result"
 # Metadata
 require "dynamics_crm/metadata/xml_document"
+require "dynamics_crm/metadata/one_to_many_relationship"
+require "dynamics_crm/metadata/relationship_metadata"
 require "dynamics_crm/metadata/entity_metadata"
 require "dynamics_crm/metadata/attribute_metadata"
+require "dynamics_crm/metadata/attribute_query_expression"
+require "dynamics_crm/metadata/filter_expression"
+require "dynamics_crm/metadata/properties_expression"
+require "dynamics_crm/metadata/entity_query_expression"
 require "dynamics_crm/metadata/retrieve_all_entities_response"
 require "dynamics_crm/metadata/retrieve_entity_response"
 require "dynamics_crm/metadata/retrieve_attribute_response"
+require "dynamics_crm/metadata/retrieve_metadata_changes_response"
 # Model
 require "dynamics_crm/model/entity"
 require "dynamics_crm/model/opportunity"
+# Fetch XML
+require "dynamics_crm/fetch_xml/entity"
+require 'dynamics_crm/fetch_xml/link_entity'
+require "dynamics_crm/fetch_xml/builder"
 # Client
 require "dynamics_crm/client"
 
@@ -33,8 +45,11 @@ require 'base64'
 require "rexml/document"
 require 'mimemagic'
 require 'curl'
+require 'securerandom'
+require 'date'
+require 'cgi'
 
-module DynamicsCRM 
+module DynamicsCRM
 
   class StringUtil
     def self.underscore(str)
@@ -43,6 +58,20 @@ module DynamicsCRM
         gsub(/([a-z\d])([A-Z])/,'\1_\2').
         tr("-", "_").
         downcase
+    end
+
+    def self.valueOf(text)
+      # Convert text to actual data types.
+      value = text
+      if value == "true" || value == "false"
+        value = (value == "true")
+      elsif value =~ /^[-?]\d+$/
+        value = value.to_i
+      elsif value =~ /^[-?]\d+\.\d+$/
+        value = value.to_f
+      else
+        value
+      end
     end
   end