+++ spec/lib/client_spec.rb
@@ -5,46 +5,82 @@ describe DynamicsCRM::Client do
   let(:subject) { DynamicsCRM::Client.new(organization_name: "tinderboxdev")}
 
   describe "#authenticate" do
-    it "authenticates with username and password" do
-
-      subject.stub(:post).and_return(fixture("request_security_token_response"))
-
-      subject.authenticate('testing', 'password')
-
-      subject.instance_variable_get("@security_token0").should start_with("tMFpDJbJHcZnRVuby5cYmRbCJo2OgOFLEOrUHj+wz")
-      subject.instance_variable_get("@security_token1").should start_with("CX7BFgRnW75tE6GiuRICjeVDV+6q4KDMKLyKmKe9A8U")
-      subject.instance_variable_get("@key_identifier").should == "D3xjUG3HGaQuKyuGdTWuf6547Lo="
-    end
-
     it "raises arugment error when no parameters are passed" do
       expect { subject.authenticate() }.to raise_error(ArgumentError)
     end
 
-    # This is only method in this suite that actually sends a POST message to Dynamics.
-    # This covers the post() and fault parsing logic.
-    it "fails to authenticate with invalid credentials" do
-      begin
-        subject.authenticate('testuser@orgnam.onmicrosoft.com', 'qwerty')
-        fail("Expected Fault to be raised")
-      rescue DynamicsCRM::XML::Fault => f
-        f.code.should == "S:Sender"
-        f.subcode.should == "wst:FailedAuthentication"
-        f.reason.should == "Authentication Failure"
+    context "Online" do
+      it "authenticates with username and password" do
+
+        subject.stub(:post).and_return(fixture("request_security_token_response"))
+
+        subject.authenticate('testing', 'password')
+
+        subject.instance_variable_get("@security_token0").should start_with("tMFpDJbJHcZnRVuby5cYmRbCJo2OgOFLEOrUHj+wz")
+        subject.instance_variable_get("@security_token1").should start_with("CX7BFgRnW75tE6GiuRICjeVDV+6q4KDMKLyKmKe9A8U")
+        subject.instance_variable_get("@key_identifier").should == "D3xjUG3HGaQuKyuGdTWuf6547Lo="
+      end
+
+      # This is only method in this suite that actually sends a POST message to Dynamics.
+      # This covers the post() and fault parsing logic.
+      it "fails to authenticate with invalid credentials" do
+        begin
+          subject.authenticate('testuser@orgnam.onmicrosoft.com', 'qwerty')
+          fail("Expected Fault to be raised")
+        rescue DynamicsCRM::XML::Fault => f
+          f.code.should == "S:Sender"
+          f.subcode.should == "wst:FailedAuthentication"
+          f.reason.should == "Authentication Failure"
+        end
       end
     end
+
+    context "On-Premise" do
+      let(:subject) { DynamicsCRM::Client.new(organization_name: "psavtest", hostname: "psavtest.crm.powerobjects.net")}
+
+      it "authenticates with username and password" do
+
+        subject.stub(:post).and_return(fixture("request_security_token_response_onpremise"))
+
+        subject.authenticate('testing', 'password')
+
+        subject.instance_variable_get("@security_token0").should start_with("ydfdQsDU9ow4XhoBi+0+n+/9Z7Dvfi")
+        subject.instance_variable_get("@security_token1").should start_with("GcCk8ivhLAAPEbQI8qScynWLReTWE0AC5")
+        subject.instance_variable_get("@key_identifier").should == "_ed121435-64ea-45b0-9b15-e5769afdb746"
+
+        subject.instance_variable_get("@cert_issuer_name").strip.should start_with("SERIALNUMBER=12369287, CN=Go Daddy Secure")
+        subject.instance_variable_get("@cert_serial_number").should == "112094107365XXXXX"
+        subject.instance_variable_get("@server_secret").should == "XZwQpJKfAy00NNWU1RwdtDpVyW/nfabuCq4H38GgKrM="
+      end
+    end
+
   end
 
   describe "#retrieve" do
-    it "retrieves object by id" do
-
+    before(:each) do
       subject.stub(:post).and_return(fixture("retrieve_account_all_columns"))
+    end
 
-      result = subject.retrieve("account", "93f0325c-a592-e311-b7f3-6c3be5a8a0c8")
+    let(:result) { subject.retrieve("account", "93f0325c-a592-e311-b7f3-6c3be5a8a0c8") }
 
+    it "retrieves object by id and acts as hash" do
       result.should be_a(DynamicsCRM::Response::RetrieveResult)
       result.type.should == "account"
       result.id.should == "93f0325c-a592-e311-b7f3-6c3be5a8a0c8"
       result.name.should == "Adventure Works (sample)"
+    end
+
+    it "exposes entity object" do
+      entity = result.entity
+
+      expect(entity).to be_a(DynamicsCRM::XML::Entity)
+      expect(entity.attributes).to be_a(DynamicsCRM::XML::Attributes)
+      expect(entity.attributes.merged).to eq(false)
+      expect(entity.attributes.nothing).to be_nil
+
+      expect(entity.formatted_values).to be_a(DynamicsCRM::XML::FormattedValues)
+      expect(entity.formatted_values.merged).to eq('No')
+      expect(entity.formatted_values.nothing).to be_nil
     end
   end
 
@@ -56,6 +92,14 @@ describe DynamicsCRM::Client do
       result = subject.retrieve_multiple("account", ["name", "Equal", "Test Account"], columns=[])
 
       result.should be_a(DynamicsCRM::Response::RetrieveMultipleResult)
+
+      expect(result['EntityName']).to eq('account')
+      expect(result['MinActiveRowVersion']).to eq(-1)
+      expect(result['MoreRecords']).to eq(false)
+      expect(result['PagingCookie']).not_to be_empty
+      expect(result['TotalRecordCount']).to eq(-1)
+      expect(result['TotalRecordCountLimitExceeded']).to eq(false)
+
       result.entities.size.should == 3
       entities = result.entities
 
@@ -98,13 +142,19 @@ describe DynamicsCRM::Client do
         </fetch>
       }
 
-      result = subject.fetch(xml)
+      entity_collection = subject.fetch(xml)
+      entity_collection.should be_a(DynamicsCRM::XML::EntityCollection)
 
-      result.should be_a(DynamicsCRM::Response::ExecuteResult)
+      expect(entity_collection.entity_name).to eq('new_tinderboxdocument')
+      expect(entity_collection.min_active_row_version).to eq(-1)
+      expect(entity_collection.more_records).to eq(false)
+      expect(entity_collection.paging_cookie).not_to be_empty
+      expect(entity_collection.total_record_count).to eq(-1)
+      expect(entity_collection.total_record_count_limit_exceeded).to eq(false)
 
-      result["EntityCollection"].size.should == 3
+      expect(entity_collection.entities.size).to eq(3)
 
-      entity = result["EntityCollection"].first
+      entity = entity_collection.entities.first
       entity.id.should == "9c27cf91-ada3-e311-b64f-6c3be5a87df0"
       entity.logical_name.should == "new_tinderboxdocument"
       entity.attributes["new_tinderboxdocumentid"].should == entity.id
@@ -250,6 +300,62 @@ describe DynamicsCRM::Client do
 
         subject.disassociate("account", "7BF2E032-AD92-E311-9752-6C3BE5A87DF0", "contact_customer_accounts", contacts)
       end
+    end
+  end
+
+  describe "#retrieve_metadata_changes" do
+    it "retrieves entity metadata" do
+      subject.stub(:post).and_return(fixture("retrieve_metadata_changes_response"))
+
+      entity_filter = DynamicsCRM::Metadata::FilterExpression.new('Or')
+      entity_filter.add_condition(['SchemaName', 'Equals', 'Contact'])
+      entity_filter.add_condition(['SchemaName', 'Equals', 'Annotation'])
+      entity_filter.add_condition(['SchemaName', 'Equals', 'Incident'])
+      entity_properties = DynamicsCRM::Metadata::PropertiesExpression.new(['Attributes'])
+
+      attribute_filter = DynamicsCRM::Metadata::FilterExpression.new('And')
+      attribute_filter.add_condition(['IsCustomAttribute', 'Equals', false])
+      attribute_properties = DynamicsCRM::Metadata::PropertiesExpression.new(['LogicalName', 'AttributeType', 'AttributeOf', 'DisplayName', 'RequiredLevel'])
+      attribute_query = DynamicsCRM::Metadata::AttributeQueryExpression.new(attribute_filter, attribute_properties)
+
+      entity_query = DynamicsCRM::Metadata::EntityQueryExpression.new({
+        criteria: entity_filter,
+        properties: entity_properties,
+        attribute_query: attribute_query
+      })
+
+      result = subject.retrieve_metadata_changes(entity_query)
+      result.should be_a(DynamicsCRM::Metadata::RetrieveMetadataChangesResponse)
+      entities = result.entities
+      entities.should_not be_nil
+      entities.size.should eq(3)
+
+      attributes = entities[0].attributes
+      attributes.size.should eq(2)
+      attribute  = attributes.first
+      attribute.logical_name.should eq("contactid")
+      attribute.attribute_of.should be_empty
+      attribute.type.should eq("Lookup")
+      attribute.display_name.should eq("Contact")
+      attribute.required_level.should eq("None")
+
+      attributes = entities[1].attributes
+      attributes.size.should eq(2)
+      attribute  = attributes.first
+      attribute.logical_name.should eq("customertypecodename")
+      attribute.attribute_of.should eq("customertypecode")
+      attribute.type.should eq("Virtual")
+      attribute.display_name.should be_empty
+      attribute.required_level.should eq("None")
+
+      attributes = entities[2].attributes
+      attributes.size.should eq(2)
+      attribute = attributes.first
+      attribute.logical_name.should eq("createdonbehalfbyyominame")
+      attribute.attribute_of.should eq("createdonbehalfby")
+      attribute.type.should eq("String")
+      attribute.display_name.should be_empty
+      attribute.required_level.should eq("None")
     end
   end